requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-core==1.32.0
azure-storage-blob==12.24.0
azure-storage-file-share==12.20.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
environs==11.2.1
et_xmlfile==2.0.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jiter==0.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
numpy==2.1.3
openai==1.55.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pythainlp==5.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
